add handling of signed and unsigned integer message data

+++ src/ConsoleWindow.cpp
@@ -53,7 +53,7 @@ void ConsoleWindow::add_log(const char *fmt, ...) // IM_FMTARGS(2)
     Items.push_back(Strdup(buf));
 }
 
-void ConsoleWindow::add_data_to_log(const uint8_t *ids, const float *values, uint32_t length)
+void ConsoleWindow::add_data_to_log(const uint8_t *ids, const double *values, uint32_t length)
 {
     for (uint32_t i = 0; i < length; i++)
     {
